Adding fixes and tests for issue #2. The issues are: * Empty character strings from the db get represented as a byte-slice with all 0s. So, in the application, a check for equality of the string with "" fails.
In _column.go_ a byte slice is used to return a value.
Initial size of that byte slice is the length of a column. Initial value
in the byte slice is 0s. As an example, if a column is VARCHAR(34) then the
byte slice has 34 0s. That byte slice was returned to the client as is.
Instead we need to return a new byte slice of the data length. In this
example, for empty string, data length is 0, so we need to return a new
byte slice of length 0. That is, the byte slice needs to be truncated
using the data length.

* Count(*) queries return column type as SQL_DECFLOAT. Handle that.

The driver didn't have an option to handle DECFLOAT. Added that option
and added tests.

* Empty binary field results in a panic, handle that.

Used the option as suggested by @shrek in issue #2.