Added tests for INOUT.
Added test for various INOUT database types to make sure that INOUT
parameter in a stored procedure works correctly.

Fixed string extract. Previously, we were extracting a string to a byte
buffer using unsafe incorrectly because unsafe.Sizeof was just the size
of the string slice header. Fixed it by defining and using
extractUTF16str.

TODO: In a varbinay INOUT test, we are supposed to repeat the input value
twice. But we only get back the original value. Investigagte ...