Allow non-select statements to use sql.Query
If Query or QueryContext is used to execute a non-select statement and
no resultset was produced, then instead of throwing a driver specific
error, use sql.ErrNoRows. That way an application can use
Query/QueryContext to run a non-select statement, which is supported by
DB2's ODBC/CLI driver and for the error, application can check if the error
is sql.ErrNoRows and move on.

TODO: Needs test.