return ErrBadConn upon connection errors
The database driver is expected to return the error ErrBadConn in some scenarios as described here:
https://golang.org/src/database/sql/driver/driver.go

ErrBadConn is used the the sql package to remove the bad connection from the connection pool (and possibly retry on a new connection) as seen here: https://golang.org/src/database/sql/sql.go

If the driver does not return ErrBadConn, then in some scenarios, the sql package is unable to prune the bad connection from the connection pool. As a result the application keeps receiving errors from the sql library due to the usage of the bad connection, even though a new connection to the database works.